"You can't have it both ways.".
Dear Michael and Shan,

I respect the tradeoffs between American and British quoting styles. On the one hand, I find the British style to be more logical, closer to the syntax you would expect in a compiled language. On the other hand, and perhaps due only to my upbringing, I think the American style looks better, the way punctuation snuggles into the end fo the sentence like that, tucked neatly into the curved quotes.

But this double period simply will not do. I've applied the British style for consistency with the rest of the piece.

Your correspondent,

R.M.O.